Struct dynalgo::graph::Graph

pub struct Graph { /* private fields */ }

Implementations

impl Graph

pub fn new() -> Graph

Creates an empty graph structure

Example:

use dynalgo::graph::Graph;
let graph = Graph::new();

pub fn node_add(&mut self, name: char, value: Option<u8>) -> Result<(), String>

Adds a node to the graph structure.

Example:

use dynalgo::graph::Graph;
use std::fs::File;
use std::io::Write;

let mut graph = Graph::new();
graph.node_add('A', None);
graph.node_add('B', None);
graph.link_add('α', 'A', 'B', true, None);
graph.node_add('C', None);
graph.link_add('β', 'C', 'B', true, None);
graph.node_add('D', None);
graph.link_add('γ', 'D', 'A', true, None);
graph.link_add('δ', 'B', 'C', true, None);

let html = graph.svg_render_animation_html("node_add example");
write!(File::create("example-node_add.html").unwrap(), "{}", html);

pub fn node_add_fixed(
&mut self,
name: char,
x: i16,
y: i16,
value: Option<u8>
) -> Result<(), String>

Adds a node to the graph structure, with a fixed (x,y) position (fixed coords of the element representation in SVG context).

Example:

use dynalgo::graph::Graph;
use std::fs::File;
use std::io::Write;

let mut graph = Graph::new();
graph.node_add_fixed('A', 400, 400, None);
graph.node_add_fixed('B', 400, 100, None);
graph.node_add_fixed('C', 100, 400, None);
graph.node_add_fixed('D', 100, 100, None);
graph.node_add('E', None);
graph.link_add('α', 'E', 'A', true, None);
graph.link_add('β', 'E', 'B', true, None);
graph.link_add('γ', 'E', 'C', true, None);
graph.link_add('δ', 'E', 'D', true, None);

let html = graph.svg_render_animation_html("node_add_fixed example");
write!(File::create("example-node_add_fixed.html").unwrap(), "{}", html);

pub fn node_delete(&mut self, name: char) -> Result<(), String>

Deletes a node from the graph structure. The associated links are also deleted.

Example:

use dynalgo::graph::Graph;
use std::fs::File;
use std::io::Write;

let mut graph = Graph::new();
graph.node_add('A', None);
graph.node_add('B', None);
graph.link_add('α', 'A', 'B', true, None);
graph.node_delete('B');

let html = graph.svg_render_animation_html("node_delete example");
write!(File::create("example-node_delete.html").unwrap(), "{}", html);

pub fn link_add(
&mut self,
name: char,
from_node: char,
to_node: char,
bidirectional: bool,
value: Option<u8>
) -> Result<(), String>

Adds a link between two nodes. The link can be defined as bidirectional or not.

Example:

use dynalgo::graph::Graph;
use std::fs::File;
use std::io::Write;

let mut graph = Graph::new();
graph.node_add('A', None);
graph.node_add('B', None);
graph.link_add('α', 'A', 'B', true, Some(10));

let html = graph.svg_render_animation_html("link_add example");
write!(File::create("example-link_add.html").unwrap(), "{}", html);

pub fn link_delete(&mut self, name: char) -> Result<(), String>

Deletes a link.

Example:

use dynalgo::graph::Graph;
use std::fs::File;
use std::io::Write;

let mut graph = Graph::new();
graph.node_add('A', None);
graph.node_add('B', None);
graph.link_add('α', 'A', 'B', true, Some(10));
graph.link_delete('α');

let html = graph.svg_render_animation_html("link_delete example");
write!(File::create("example-link_delete.html").unwrap(), "{}", html);

pub fn node_value(&self, name: char) -> Result<Option<u8>, String>

Returns the optional value associated to the node.

Example:

use dynalgo::graph::Graph;

let mut graph = Graph::new();
graph.node_add('A', Some(10));
let value : Option<u8> = graph.node_value('A').unwrap();

pub fn nodes_list(&self) -> Vec<char>

Returns the list of nodes names.

Example:

use dynalgo::graph::Graph;

let mut graph = Graph::new();
graph.node_add('A', None);
graph.node_add('B', None);
let nodes : Vec<char> = graph.nodes_list();

pub fn adjacency_list(&self) -> HashMap<char, HashMap<char, (char, Option<u8>)>>

Returns an adjacency list.

Example:

use dynalgo::graph::Graph;
use std::collections::HashMap;

let mut graph = Graph::new();
graph.node_add('A', None);
graph.node_add('B', None);
graph.link_add('α', 'A', 'B', true, Some(10));

let adjacency : HashMap<char, HashMap<char, (char, Option<u8>)>> = graph.adjacency_list();
for (node_from, adjacent) in adjacency {
    for (node_to, link) in adjacent {
        println!("Can go from node {} to node {}", node_from, node_to);
        match link {
            (link_name, Some(value)) => println!("Link {} has value {}", link_name, value),
            (link_name, None) => println!("Link is {}", link_name),
        }
    }
}

pub fn nodes_exchange(&mut self, name_1: char, name_2: char) -> Result<(), String>

Exchanges two nodes in the graph structure.

Example:

use dynalgo::graph::Graph;
use std::fs::File;
use std::io::Write;

let mut graph = Graph::new();
graph.svg_param_display_link_label(true);
graph.node_add('A', None);
graph.node_add('B', None);
graph.node_add('C', None);
graph.link_add('α', 'A', 'B', true, None);
graph.link_add('β', 'B', 'C', true, None);
graph.link_add('γ', 'C', 'A', true, None);
graph.nodes_exchange('A', 'B');

let html = graph.svg_render_animation_html("nodes_exchange example");
write!(File::create("example-nodes_exchange.html").unwrap(), "{}", html);

pub fn dyna_from(&mut self, graph_config: String) -> Result<(), String>

Imports a graph structure from a formatted String

Example:

use dynalgo::graph::Graph;
use std::fs::File;
use std::io::Write;

let mut graph = Graph::new();
let dyna = String::from(
       "N A _ _ _
        N B _ _ _
        N C _ _ _
        L a A B true 1
        L b A C true 2
        L c B C true 3"
);
graph.dyna_from(dyna);

let html = graph.svg_render_animation_html("dyna_from example");
write!(File::create("example-dyna_from.html").unwrap(), "{}", html);

pub fn dyna_to(&self) -> String

Exports a graph structure to a formatted String

Example:

use dynalgo::graph::Graph;
use std::fs::File;
use std::io::Write;

let mut graph = Graph::new();
graph.node_add('A', None);
graph.node_add('B', Some(2));
graph.link_add('α', 'A', 'B', true, Some(1));
let dyna = graph.dyna_to();   
    
let mut dyna_file = File::create("example-dyna_to.dyna").unwrap();
write!(dyna_file, "{}", dyna);

pub fn svg_automatic_animation(&mut self, automatic: bool)

Activates or desactivates the svg_automatic_animation option (affects only SVG rendering). When this option is activated, each change in the structure of the graph, and each change in the property of the representation of the node or of the link causes a graphic animation (the animations occur one after the other). If this option is desactivated, then all the pending animations occur during the same period (i.e. not one after the other) when you manually call the svg_animate function.

Example:

use dynalgo::graph::Graph;
use std::fs::File;
use std::io::Write;

let mut graph = Graph::new();
graph.node_add('A', None);
graph.node_add('B', None);
graph.link_add('α', 'A', 'B', true, Some(10));
graph.svg_automatic_animation(false);
graph.node_add_fixed('C', 0, 60, None);
graph.node_add_fixed('D', 0, -60, None);
graph.link_add('β', 'A', 'C', true, Some(20));
graph.link_add('γ', 'D', 'C', true, Some(30));
graph.svg_animate(5000);
graph.svg_automatic_animation(true);
graph.svg_layout();

let html = graph.svg_render_animation_html("svg_automatic_animation example");
write!(File::create("example-svg_automatic_animation.html").unwrap(), "{}", html);

pub fn svg_animate(&mut self, duration_ms: u32)

Creates SVG animations with SMIL language (affects only SVG rendering). Each change in the structure of the graph, and each change in the property of the representation of the node or of the link causes an animation. The animation show the evolution between the last represented state and the current state. The value of the duration_ms parameter can not be less than 1 ms.

Example:

use dynalgo::graph::Graph;
use std::fs::File;
use std::io::Write;

let mut graph = Graph::new();
graph.node_add_fixed('A', 0, 0, None);
graph.node_add_fixed('B', 100, 0, None);
graph.node_add_fixed('C', 50, -50, None);
graph.link_add('α', 'A', 'B', true, None);
graph.svg_node_move('A', 50, 0);
graph.svg_node_move('A', 50, 50);
graph.svg_node_move('A', 0, 50);
graph.svg_automatic_animation(false);
graph.svg_node_move('B', 150, 0);
graph.svg_node_move('B', 150, 50);
graph.svg_node_move('B', 100, 50);
graph.svg_animate(1000);
graph.svg_automatic_animation(true);
graph.nodes_exchange('A', 'B');
graph.svg_automatic_animation(false);
graph.svg_node_color('A', 255, 0, 0);
graph.svg_node_color('B', 0, 255, 0);
graph.nodes_exchange('A', 'B');
graph.nodes_exchange('A', 'B');
graph.svg_node_color('C', 0, 0, 255);
graph.svg_animate(1000);

let html = graph.svg_render_animation_html("svg_animate example");
write!(File::create("example-svg_animate.html").unwrap(), "{}", html);

pub fn svg_automatic_layout(&mut self, automatic: bool)

Activates or desactivates the svg_automatic_layout option (affects only SVG rendering). Activating this option is revelant only if the svg_automatic_animation option is activated. When this option is activated, the nodes are layouted as they are created. When this option is desactivated, you must explicitly call the svg_layout or svg_layout_nodes function to layout the created nodes.

Example:

use dynalgo::graph::Graph;
use std::fs::File;
use std::io::Write;

let mut graph = Graph::new();
graph.svg_automatic_layout(false);
graph.node_add('A', None);
graph.node_add('B', None);
graph.link_add('α', 'A', 'B', true, Some(10));
graph.svg_layout();

let html = graph.svg_render_animation_html("svg_automatic_layout example");
write!(File::create("example-svg_automatic_layout.html").unwrap(), "{}", html);

pub fn svg_layout_nodes(&mut self, nodes_names: Vec<char>) -> Result<(), String>

Layouts the listed nodes (affects only SVG rendering). Calling this function is revelant only if the option svg_automatic_layout is desactivated.

Example:

use dynalgo::graph::Graph;
use std::fs::File;
use std::io::Write;

let mut graph = Graph::new();
graph.node_add('A', None);
graph.node_add('B', None);
graph.link_add('α', 'A', 'B', true, None);
graph.svg_automatic_layout(false);
graph.node_add('C', None);
graph.node_add('D', None);
graph.link_add('β', 'A', 'C', true, None);
graph.link_add('γ', 'B', 'D', true, None);
graph.svg_layout_nodes(vec!['C', 'D']);

let html = graph.svg_render_animation_html("svg_layout_nodes example");
write!(File::create("example-svg_layout_nodes.html").unwrap(), "{}", html);

pub fn svg_layout(&mut self)

Layouts all the nodes (affects only SVG rendering). Calling this function is revelant only if the svg_automatic_layout option is desactivated.

Example:

use dynalgo::graph::Graph;
use std::fs::File;
use std::io::Write;

let mut graph = Graph::new();
graph.svg_automatic_layout(false);
graph.node_add('A', None);
graph.node_add('B', None);
graph.link_add('α', 'A', 'B', true, None);
graph.node_add('C', None);
graph.node_add('D', None);
graph.link_add('β', 'A', 'C', true, None);
graph.link_add('γ', 'B', 'D', true, None);
graph.svg_layout();

let html = graph.svg_render_animation_html("svg_layout example");
write!(File::create("example-svg_layout.html").unwrap(), "{}", html);

pub fn svg_render_animation_html(&self, title: &str) -> String

Exports the animations in SVG format into a HTML page. The HTML page contains a little javascript so that the animation can be paused and resumed by clickig on the image.

Example:

use dynalgo::graph::Graph;
use std::fs::File;
use std::io::Write;

let mut graph = Graph::new();
graph.node_add('A', None);
graph.node_add('B', None);
graph.node_add('C', None);
graph.link_add('α', 'A', 'B', true, Some(10));
graph.link_add('β', 'B', 'C', true, Some(20));
graph.link_add('γ', 'C', 'A', true, Some(30));

let html_file_content = graph.svg_render_animation_html("svg_render_animation_html example");       
let mut html_file = File::create("example-svg_render_animation_html.html").unwrap();
write!(html_file, "{}", html_file_content);

pub fn svg_node_position(&self, name: char) -> Result<(i32, i32), String>

Returns the current x,y coords of the node in the SVG graphic context.

Example:

use dynalgo::graph::Graph;
let mut graph = Graph::new();
graph.node_add('A', None);
let (x,y) : (i32,i32) = graph.svg_node_position('A').unwrap();

pub fn svg_node_selected(
&mut self,
name: char,
selected: bool
) -> Result<(), String>

Changes the selected status of a node (affects only SVG rendering). The selected status affects the stroke color of the node.

Example:

use dynalgo::graph::Graph;
use std::fs::File;
use std::io::Write;

let mut graph = Graph::new();
graph.svg_automatic_animation(false);
graph.node_add('A', None);
graph.node_add('B', None);
graph.node_add('C', None);
graph.link_add('α', 'A', 'B', true, Some(10));
graph.link_add('β', 'B', 'C', true, Some(20));
graph.link_add('γ', 'C', 'A', true, Some(30));
graph.svg_automatic_animation(true);
graph.svg_layout();
graph.svg_node_selected('A', true);
graph.svg_node_selected('B', true);
graph.svg_node_selected('C', true);

let html = graph.svg_render_animation_html("svg_node_selected example");
write!(File::create("example-svg_node_selected.html").unwrap(), "{}", html);

pub fn svg_link_selected(
&mut self,
name: char,
selected: bool
) -> Result<(), String>

Changes the selected status of a link (affects only SVG rendering). The selected status affects the stroke color of the link.

Example:

use dynalgo::graph::Graph;
use std::fs::File;
use std::io::Write;

let mut graph = Graph::new();
graph.svg_automatic_animation(false);
graph.node_add('A', None);
graph.node_add('B', None);
graph.node_add('C', None);
graph.link_add('α', 'A', 'B', true, Some(10));
graph.link_add('β', 'B', 'C', true, Some(20));
graph.link_add('γ', 'C', 'A', true, Some(30));
graph.svg_automatic_animation(true);
graph.svg_layout();
graph.svg_link_selected('α', true);
graph.svg_link_selected('β', true);
graph.svg_link_selected('γ', true);

let html = graph.svg_render_animation_html("svg_link_selected example");
write!(File::create("example-svg_link_selected.html").unwrap(), "{}", html);

pub fn svg_node_move(&mut self, name: char, cx: i32, cy: i32) -> Result<(), String>

Changes the x,y coords of the node SVG representation (affects only SVG rendering).

Example:

use dynalgo::graph::Graph;
use std::fs::File;
use std::io::Write;

let mut graph = Graph::new();
graph.node_add_fixed('A', -50, 0, None);
graph.node_add_fixed('C', 450, 400, None);
graph.node_add('B', None);
graph.link_add('α', 'A', 'B', true, Some(10));
graph.link_add('β', 'B', 'C', true, Some(20));
graph.svg_node_move('B', 0, 0);
graph.svg_node_move('B', 0, 400);
graph.svg_node_move('B', 400, 400);
graph.svg_node_move('B', 400, 0);
graph.svg_node_move('B', 0, 0);

let html = graph.svg_render_animation_html("svg_node_move example");
write!(File::create("example-svg_node_move.html").unwrap(), "{}", html);

pub fn svg_node_color(
&mut self,
name: char,
red: u8,
green: u8,
blue: u8
) -> Result<(), String>

Changes the fill color of the node SVG representation (affects only SVG rendering).

Example:

use dynalgo::graph::Graph;
use std::fs::File;
use std::io::Write;

let mut graph = Graph::new();
graph.svg_automatic_animation(false);
graph.node_add('A', None);
graph.node_add('B', None);
graph.node_add('C', None);
graph.link_add('α', 'A', 'B', true, Some(10));
graph.link_add('β', 'B', 'C', true, Some(20));
graph.link_add('γ', 'C', 'A', true, Some(30));
graph.svg_automatic_animation(true);
graph.svg_layout();
graph.svg_node_color('A', 255, 0, 0);
graph.svg_node_color('B', 0, 255, 0);
graph.svg_node_color('C', 0, 0, 255);
graph.svg_node_color('A', 0, 255, 0);
graph.svg_node_color('B', 0, 0, 255);
graph.svg_node_color('C', 255, 0, 0);

let html = graph.svg_render_animation_html("svg_node_color example");
write!(File::create("example-svg_node_color.html").unwrap(), "{}", html);

pub fn svg_param_duration_add(&mut self, duration_ms: u32)

Changes the value of the duration_add parameter (affects only SVG rendering). This parameter affects the duration of the animation when a node or a link is added. Default value is 1000 ms

pub fn svg_param_duration_delete(&mut self, duration_ms: u32)

Changes the value of the duration_delete parameter (affects only SVG rendering). This parameter affects the duration of the animation when a node or a link is deleted. Default value is 1000 ms

pub fn svg_param_duration_move(&mut self, duration_ms: u32)

Changes the value of the duration_move parameter (affects only SVG rendering). This parameter affects the duration of the animation when a node is moved. Default value is 1000 ms

pub fn svg_param_duration_select(&mut self, duration_ms: u32)

Changes the value of the duration_select parameter (affects only SVG rendering). This parameter affects the duration of the animation when a node or a link is selected or deselected. Default value is 1000 ms

pub fn svg_param_duration_color(&mut self, duration_ms: u32)

Changes the value of the duration_color parameter (affects only SVG rendering). This parameter affects the duration of the animation when a node is colored. Default value is 1000 ms

pub fn svg_param_color_tag_created(&mut self, (red, green, blue): (u8, u8, u8))

Changes the value of the color_tag_created parameter (affects only SVG rendering). This parameter affects the stroke color of a node or a link when just created. Default rgb value is (0, 0, 255)

pub fn svg_param_color_tag_selected(&mut self, (red, green, blue): (u8, u8, u8))

Changes the value of the color_tag_selected parameter (affects only SVG rendering). This parameter affects the stroke color of a node or a link when selected. Default rgb value is (191, 255, 0)

pub fn svg_param_color_tag_deleted(&mut self, (red, green, blue): (u8, u8, u8))

Changes the value of the color_tag_deleted parameter (affects only SVG rendering). This parameter affects the color of a node or a link when deleted. Default rgb value is (255, 0, 0)

pub fn svg_param_color_node_fill(&mut self, (red, green, blue): (u8, u8, u8))

Changes the value of the color_node_fill parameter (affects only SVG rendering). This parameter affects the default fill color of a node. Default rgb value is (255, 255, 255)

pub fn svg_param_color_node_stroke(&mut self, (red, green, blue): (u8, u8, u8))

Changes the value of the color_node_stroke parameter (affects only SVG rendering). This parameter affects the default stroke color of a node. Default rgb value is (128, 139, 150)

pub fn svg_param_color_link_stroke(&mut self, (red, green, blue): (u8, u8, u8))

Changes the value of the color_link_stroke parameter (affects only SVG rendering). This parameter affects the stroke color of a link. Default rgb value is (128, 139, 150)

pub fn svg_param_color_text(&mut self, (red, green, blue): (u8, u8, u8))

Changes the value of the color_text parameter (affects only SVG rendering). This parameter affects the font color. Default rgb value is (0, 0, 0)

pub fn svg_param_display_node_label(&mut self, display: bool)

Changes the value of the display_node_label parameter (affects only SVG rendering). This parameter indicates whether the node label is displayed or not. Default value is True

pub fn svg_param_display_node_value(&mut self, display: bool)

Changes the value of the display_node_value parameter (affects only SVG rendering). This parameter indicates whether the node value is displayed or not. Default value is False

pub fn svg_param_display_link_label(&mut self, display: bool)

Changes the value of the display_link_label parameter (affects only SVG rendering). This parameter indicates whether the link label is displayed or not. Default value is False

pub fn svg_param_display_link_value(&mut self, display: bool)

Changes the value of the display_link_value parameter (affects only SVG rendering). This parameter indicates whether the link value is displayed or not. Default value is True

pub fn svg_param_stroke_width_node(&mut self, width: u32)

Changes the value of the stroke_width_node parameter (affects only SVG rendering). This parameter affects the stroke width of the nodes. Default value is 2

pub fn svg_param_stroke_width_link(&mut self, width: u32)

Changes the value of the stroke_width_link parameter (affects only SVG rendering). This parameter affects the stroke width of the links. Default value is 2

pub fn svg_param_radius_node(&mut self, radius: u32)

Changes the value of the radius_node parameter (affects only SVG rendering). This parameter affects the radius of the nodes. Default value is 20